requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
docopt==0.6.2
Flask==3.1.2
Flask-Minify==0.50
htmlminf==0.1.13
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsmin==3.0.1
lesscpy==0.15.1
Markdown==3.10.1
MarkupSafe==3.0.3
pipreqs==0.4.13
ply==3.11
Pygments==2.19.2
python-dotenv==1.2.1
python-frontmatter==1.1.0
PyYAML==6.0.3
rcssmin==1.2.2
requests==2.32.5
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.5
xxhash==3.6.0
yarg==0.1.10
